Refactor database connections to use shared pool across all services
