+++ backend/admin_pricing_dashboard.py
@@ -111,8 +111,13 @@ class AdminPricingDashboard:
     async def _get_pricing_history(self) -> List[Dict[str, Any]]:
         """Get pricing history for the last 30 days"""
         try:
-            conn = await asyncpg.connect(self.database_url)
-            try:
+            import db
+            pool = db.get_db_pool()
+            if not pool:
+                logger.warning("Shared database pool not available for pricing dashboard")
+                return []
+                
+            async with pool.acquire() as conn:
                 rows = await conn.fetch("""
                     SELECT service_name, old_price, new_price, reasoning, changed_at
                     FROM pricing_history
@@ -135,9 +140,6 @@ class AdminPricingDashboard:
                 
                 return history
                 
-            finally:
-                await conn.close()
-                
         except Exception as e:
             logger.error(f"Pricing history error: {e}")
             return []
@@ -145,8 +147,11 @@ class AdminPricingDashboard:
     async def _get_demand_analytics(self) -> Dict[str, Any]:
         """Get demand analytics for pricing decisions"""
         try:
-            conn = await asyncpg.connect(self.database_url)
-            try:
+            import db
+            pool = db.get_db_pool()
+            if not pool:
+                return None
+            async with pool.acquire() as conn:
                 # Get session counts by day for last 30 days
                 daily_rows = await conn.fetch("""
                     SELECT 
@@ -206,9 +211,6 @@ class AdminPricingDashboard:
                     "peak_hours": sorted(hourly_demand, key=lambda x: x["sessions"], reverse=True)[:3]
                 }
                 
-            finally:
-                await conn.close()
-                
         except Exception as e:
             logger.error(f"Demand analytics error: {e}")
             return {
@@ -224,8 +226,11 @@ class AdminPricingDashboard:
     async def _calculate_revenue_impact(self) -> Dict[str, Any]:
         """Calculate revenue impact of pricing changes"""
         try:
-            conn = await asyncpg.connect(self.database_url)
-            try:
+            import db
+            pool = db.get_db_pool()
+            if not pool:
+                return None
+            async with pool.acquire() as conn:
                 # Get revenue for last 30 days
                 revenue_row = await conn.fetchrow("""
                     SELECT 
@@ -292,9 +297,6 @@ class AdminPricingDashboard:
                     }
                 }
                 
-            finally:
-                await conn.close()
-                
         except Exception as e:
             logger.error(f"Revenue impact calculation error: {e}")
             return {
@@ -357,8 +359,11 @@ class AdminPricingDashboard:
                     "message": "Dynamic pricing system not available"
                 }
             
-            conn = await asyncpg.connect(self.database_url)
-            try:
+            import db
+            pool = db.get_db_pool()
+            if not pool:
+                return None
+            async with pool.acquire() as conn:
                 # Store pricing override
                 await conn.execute("""
                     INSERT INTO pricing_overrides 
@@ -390,9 +395,6 @@ class AdminPricingDashboard:
                     "override_price": override_request.override_price,
                     "expires_at": (datetime.now() + timedelta(hours=override_request.duration_hours)).isoformat()
                 }
-                
-            finally:
-                await conn.close()
                 
         except Exception as e:
             logger.error(f"Pricing override error: {e}")
